move components to variant page folder
