+++ src/component/variantPage/BasicInfo.css
