+++ src/component/variantPage/BasicInfo.tsx
