+++ src/component/variantPage/CheckBox.css
@@ -24,7 +24,7 @@
 }
 
 .custom-checkbox label a {
-    padding-left: 0px !important;
+    padding-left: 0.2rem !important;
     text-align: left;
     color: black;
 }
