+++ src/component/variantPage/SideBar.css
