+++ src/component/variantPage/TranscriptSummaryTable.css
