+++ src/component/variantPage/TranscriptSummaryTable.tsx
