--- src/component/variantPage/variantComponentHeader.css
@@ -1,6 +1,0 @@
-.header {
-    border-bottom: 2px solid #E0E0E0;
-    font-size: 1.2rem;
-    font-weight: 600;
-    color: #0355A6;
-}
